+++ config.go
@@ -45,6 +45,7 @@ type Config struct {
 	LightstepReconnectPeriod      string    `yaml:"lightstep_reconnect_period"`
 	MetricMaxLength               int       `yaml:"metric_max_length"`
 	NumReaders                    int       `yaml:"num_readers"`
+	NumSpanWorkers                int       `yaml:"num_span_workers"`
 	NumWorkers                    int       `yaml:"num_workers"`
 	OmitEmptyHostname             bool      `yaml:"omit_empty_hostname"`
 	Percentiles                   []float64 `yaml:"percentiles"`
@@ -53,6 +54,7 @@ type Config struct {
 	SignalfxAPIKey                string    `yaml:"signalfx_api_key"`
 	SignalfxEndpointBase          string    `yaml:"signalfx_endpoint_base"`
 	SignalfxHostnameTag           string    `yaml:"signalfx_hostname_tag"`
+	SpanChannelCapacity           int       `yaml:"span_channel_capacity"`
 	SsfBufferSize                 int       `yaml:"ssf_buffer_size"`
 	SsfListenAddresses            []string  `yaml:"ssf_listen_addresses"`
 	StatsAddress                  string    `yaml:"stats_address"`