+++ trace/client_test.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
 	"github.com/DataDog/datadog-go/statsd"
 	"github.com/stretchr/testify/assert"
@@ -18,6 +17,38 @@ import (
 	"github.com/stripe/veneur/protocol"
 	"github.com/stripe/veneur/ssf"
 )
+
+func mustRecord(t *testing.T, client *Client, tr *Trace) (retries int) {
+	for {
+		err := tr.ClientRecord(client, "", map[string]string{})
+		if err != ErrWouldBlock {
+			assert.NoError(t, err)
+			return
+		}
+		t.Log("retrying record")
+		retries++
+	}
+}
+
+func mustFlush(t *testing.T, client *Client) (retries int) {
+	anyBlockage := func(err *FlushError) bool {
+		for _, subErr := range err.Errors {
+			if subErr != ErrWouldBlock {
+				return false
+			}
+		}
+		return true
+	}
+	for {
+		err := Flush(client)
+		if err == nil || !anyBlockage(err.(*FlushError)) {
+			require.NoError(t, err)
+			return
+		}
+		t.Log("retrying flush")
+		retries++
+	}
+}
 
 func TestNoClient(t *testing.T) {
 	err := Record(nil, nil, nil)
@@ -85,8 +116,7 @@ func TestUNIX(t *testing.T) {
 		name := fmt.Sprintf("Testing-%d", i)
 		tr := StartTrace(name)
 		tr.Sent = sentCh
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 	}
 	for i := 0; i < 4; i++ {
 		assert.NoError(t, <-sentCh)
@@ -118,6 +148,7 @@ func TestUNIXBuffered(t *testing.T) {
 
 	client, err := NewClient((&url.URL{Scheme: "unix", Path: sockName}).String(),
 		Capacity(4),
+		ParallelBackends(1),
 		Buffered)
 	require.NoError(t, err)
 	defer client.Close()
@@ -127,70 +158,20 @@ func TestUNIXBuffered(t *testing.T) {
 		name := fmt.Sprintf("Testing-%d", i)
 		tr := StartTrace(name)
 		tr.Sent = sentCh
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 	}
 	for i := 0; i < 4; i++ {
 		assert.NoError(t, <-sentCh)
 	}
 	assert.Equal(t, 0, len(outPkg), "Should not have sent any packets yet")
-	err = Flush(client)
-	assert.NoError(t, err)
+
+	mustFlush(t, client)
 	for i := 0; i < 4; i++ {
 		<-outPkg
 	}
 }
 
-func TestUNIXBufferedFlushing(t *testing.T) {
-	dir, err := ioutil.TempDir("", "test_unix")
-	require.NoError(t, err)
-	defer os.RemoveAll(dir)
-
-	sockName := filepath.Join(dir, "sock")
-	laddr, err := net.ResolveUnixAddr("unix", sockName)
-	require.NoError(t, err)
-
-	outPkg := make(chan *ssf.SSFSpan, 4)
-	cleanup := serveUNIX(t, laddr, func(in net.Conn) {
-		for {
-			pkg, err := protocol.ReadSSF(in)
-			if err == io.EOF {
-				return
-			}
-			assert.NoError(t, err)
-			outPkg <- pkg
-		}
-	})
-	defer cleanup()
-	flushChan := make(chan time.Time)
-
-	client, err := NewClient((&url.URL{Scheme: "unix", Path: sockName}).String(),
-		Capacity(4),
-		Buffered,
-		FlushChannel(flushChan, func() {}))
-	require.NoError(t, err)
-	defer client.Close()
-
-	sentCh := make(chan error)
-	for i := 0; i < 4; i++ {
-		name := fmt.Sprintf("Testing-%d", i)
-		tr := StartTrace(name)
-		tr.Sent = sentCh
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
-	}
-	for i := 0; i < 4; i++ {
-		assert.NoError(t, <-sentCh)
-	}
-	assert.Equal(t, 0, len(outPkg), "Should not have sent any packets yet")
-	flushChan <- time.Now()
-	assert.NoError(t, err)
-	for i := 0; i < 4; i++ {
-		<-outPkg
-	}
-}
-
-func serveUNIX(t *testing.T, laddr *net.UnixAddr, onconnect func(conn net.Conn)) (cleanup func() error) {
+func serveUNIX(t testing.TB, laddr *net.UnixAddr, onconnect func(conn net.Conn)) (cleanup func() error) {
 	srv, err := net.ListenUnix(laddr.Network(), laddr)
 	require.NoError(t, err)
 	cleanup = srv.Close
@@ -234,8 +215,7 @@ func TestFailingUDP(t *testing.T) {
 		name := fmt.Sprintf("Testing-%d", i)
 		tr := StartTrace(name)
 		tr.Sent = sentCh
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 	}
 	for i := 0; i < 4; i++ {
 		// Linux reports an error when sending to a
@@ -272,7 +252,10 @@ func TestReconnectUNIX(t *testing.T) {
 	})
 	defer cleanup()
 
-	client, err := NewClient((&url.URL{Scheme: "unix", Path: sockName}).String(), Capacity(4))
+	client, err := NewClient((&url.URL{Scheme: "unix", Path: sockName}).String(),
+		Capacity(4),
+		ParallelBackends(1),
+	)
 	require.NoError(t, err)
 	defer client.Close()
 
@@ -282,8 +265,7 @@ func TestReconnectUNIX(t *testing.T) {
 		tr := StartTrace(name)
 		tr.Sent = sentCh
 		t.Logf("submitting span")
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 
 		// A span was successfully received by the server:
 		assert.NoError(t, <-sentCh)
@@ -294,8 +276,7 @@ func TestReconnectUNIX(t *testing.T) {
 		tr := StartTrace(name)
 		tr.Sent = sentCh
 		t.Logf("submitting span")
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 
 		// Since reconnections throw away the span, nothing
 		// was received:
@@ -306,8 +287,7 @@ func TestReconnectUNIX(t *testing.T) {
 		tr := StartTrace(name)
 		tr.Sent = sentCh
 		t.Logf("submitting span")
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 
 		// A span was successfully received by the server:
 		assert.NoError(t, <-sentCh)
@@ -342,6 +322,7 @@ func TestReconnectBufferedUNIX(t *testing.T) {
 
 	client, err := NewClient((&url.URL{Scheme: "unix", Path: sockName}).String(),
 		Capacity(4),
+		ParallelBackends(1),
 		Buffered)
 	require.NoError(t, err)
 	defer client.Close()
@@ -352,12 +333,11 @@ func TestReconnectBufferedUNIX(t *testing.T) {
 		tr := StartTrace(name)
 		tr.Sent = sentCh
 		t.Logf("submitting span")
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 
 		assert.NoError(t, <-sentCh, "at %q", name)
 		t.Logf("Flushing")
-		go assert.NoError(t, Flush(client))
+		go mustFlush(t, client)
 		// A span was successfully received by the server:
 		<-outPkg
 	}
@@ -366,8 +346,7 @@ func TestReconnectBufferedUNIX(t *testing.T) {
 		tr := StartTrace(name)
 		tr.Sent = sentCh
 		t.Logf("submitting span")
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 
 		assert.NoError(t, <-sentCh)
 		t.Logf("Flushing")
@@ -380,13 +359,12 @@ func TestReconnectBufferedUNIX(t *testing.T) {
 		tr := StartTrace(name)
 		tr.Sent = sentCh
 		t.Logf("submitting span")
-		err = tr.ClientRecord(client, name, map[string]string{})
-		assert.NoError(t, err)
+		mustRecord(t, client, tr)
 
 		assert.NoError(t, <-sentCh, "at %q", name)
 
 		t.Logf("Flushing")
-		go assert.NoError(t, Flush(client))
+		go mustFlush(t, client)
 		// A span was successfully received by the server:
 		<-outPkg
 	}
@@ -449,8 +427,9 @@ func TestInternalBackend(t *testing.T) {
 }
 
 type successTestBackend struct {
-	t     *testing.T
-	block chan func()
+	t         *testing.T
+	block     chan chan struct{}
+	flushChan chan chan<- error
 }
 
 func (tb *successTestBackend) Close() error {
@@ -461,8 +440,8 @@ func (tb *successTestBackend) Close() error {
 func (tb *successTestBackend) SendSync(ctx context.Context, span *ssf.SSFSpan) error {
 	tb.t.Logf("Sending span")
 	select {
-	case fun := <-tb.block:
-		fun()
+	case block := <-tb.block:
+		<-block
 	default:
 	}
 	return nil
@@ -471,57 +450,49 @@ func (tb *successTestBackend) SendSync(ctx context.Context, span *ssf.SSFSpan) e
 func (tb *successTestBackend) FlushSync(ctx context.Context) error {
 	tb.t.Logf("flushing")
 	select {
-	case fun := <-tb.block:
-		fun()
+	case block := <-tb.block:
+		<-block
 	default:
 	}
 	return nil
 }
 
+func (tb *successTestBackend) FlushChan() chan chan<- error {
+	return tb.flushChan
+}
+
 func TestDropStatistics(t *testing.T) {
-	blockNext := make(chan func(), 1)
-	tb := successTestBackend{t, blockNext}
+	blockNext := make(chan chan struct{}, 1)
+	flushChan := make(chan chan<- error)
+	defer close(flushChan)
+	tb := successTestBackend{t: t, block: blockNext, flushChan: flushChan}
 
 	// Make a client that blocks if nothing listens:
 	cl, err := NewBackendClient(&tb, Capacity(0))
 	require.NoError(t, err)
 
-	somespan := func() error {
-		tr := StartTrace("hi there")
-		return tr.ClientRecord(cl, "hi there", map[string]string{})
-	}
-
-	for err = Flush(cl); err != nil; err = Flush(cl) {
-		// Wait until the client is ready (this can take a few
-		// cycles, as the goroutine to read ops must spin up)
-	}
 	// reset client stats:
 	stats, err := statsd.NewBuffered("127.0.0.1:8200", 4096)
 	require.NoError(t, err)
 	SendClientStatistics(cl, stats, nil)
 
 	// Actually test the client:
-	err = Flush(cl)
+	flushRetries := mustFlush(t, cl)
 	assert.NoError(t, err, "Flushing an empty client should succeed")
 	assert.Equal(t, int64(1), cl.successfulFlushes, "successful flushes")
 
 	done := make(chan struct{})
-	blockNext <- func() {
-		<-done
-	}
-	err = somespan()
-	assert.NoError(t, err, "Sending an expected metric should succeed")
+	blockNext <- done
+	retries := mustRecord(t, cl, StartTrace("hi there"))
 	assert.Equal(t, int64(1), cl.successfulRecords, "successful records")
 
-	err = somespan()
+	err = StartTrace("hi there").ClientRecord(cl, "", map[string]string{})
 	assert.Error(t, err)
 	assert.Equal(t, ErrWouldBlock, err, "Expected to report a blocked channel")
-	assert.Equal(t, int64(1), cl.failedRecords, "failed records")
+	assert.Equal(t, int64(1+retries), cl.failedRecords, "failed records")
 
 	err = Flush(cl)
-	assert.Error(t, err)
-	assert.Equal(t, ErrWouldBlock, err, "Expected to report a blocked channel")
-	assert.Equal(t, int64(1), cl.failedFlushes, "failed flushes")
+	assert.Equal(t, int64(1+flushRetries), cl.failedFlushes, "failed flushes")
 	close(done)
 	close(blockNext)
 }